+++ src/nmea/sentences/mod.rs
@@ -4,6 +4,8 @@
 mod ack;
 #[cfg(feature = "aam")]
 mod aam;
+#[cfg(feature = "acn")]
+mod acn;
 #[cfg(feature = "apb")]
 mod apb;
 #[cfg(feature = "bec")]
@@ -121,6 +123,8 @@ mod zda;
 pub use ack::*;
 #[cfg(feature = "aam")]
 pub use aam::*;
+#[cfg(feature = "acn")]
+pub use acn::*;
 #[cfg(feature = "apb")]
 pub use apb::*;
 #[cfg(feature = "bec")]